changed auto procedure names in lazdev
duplicated google module with lazdev and replaced original with
automated version
