+++ modules/module.google/config.json
@@ -1,48 +1,60 @@
 {
-
-    "id":"google",
-
-    "name":"Google",
-
-    "version":"2.2",
-    
-    "type":"command",
-    
-    "permission":"anyone",
-    
-    "commands":{
-    
-    	"search":{
-			"aliases":["gs"],
-			"procedure":"doSearch",
-			"args":["text"],
-			"help":{
-	    		"id":"search",
-	    		"title":"Google quick-search",
-	            "text":"Perform a quick google search without ever leaving the comfort of your discord",
-	            "example":"%PREFIX%%COMMAND% <stuff to google>"
-	    	}
-		},
-    	"translate":{
-			"aliases":["gt"],
-			"procedure":"doTranslate",
-			"args":["id", "lang", "num"],
-			"help":{
-	        	"id":"translate",
-	        	"title":"Google Translation help",
-	        	"text":"Translate a specified number of users last messages into specified language.",
-	        	"example":"%PREFIX%%COMMAND% @shittybill fr 4"
-	    	}
-		}
-    
+  "id": "google",
+  "name": "Google Tools",
+  "version": "0.1",
+  "type": "command",
+  "permission": "anyone",
+  "commands": {
+    "search": {
+      "aliases": [
+        "gs"
+      ],
+      "procedure": "doSearch",
+      "args": [
+        "text"
+      ],
+      "help": {
+        "id": "search",
+        "title": "Quick-search Google",
+        "text": "...Do you feel lucky?",
+        "example": "%PREFIX%%COMMAND% <stuff to search>"
+      },
+      "subcommands": {
+        "site": {
+          "aliases": [
+            "s"
+          ],
+          "procedure": "doSearchSite",
+          "args": [
+            "id",
+            "text"
+          ],
+          "help": {
+            "id": "site",
+            "title": "Quick-search a site from Google ",
+            "text": "...Do you feel lucky?",
+            "example": "%PREFIX%%SUBCMD% <site.com> <stuff to search>"
+          }
+        }
+      }
     },
-        
-    "queries":{
-    	
-    	"GET_CHANNEL_SETTINGS":"SELECT * FROM `channel` WHERE `channelID` = ?",
-    	"GET_SERVER_SETTINGS":"SELECT * FROM `channel` WHERE `serverID` = ?",
-    	"SET_SETTINGS":"INSERT INTO `channel` (`channelID`, `channelName`, `serverID`, `serverName`, `region`, `memberCount`, `spreadsheet`, `webhook`, `modrole`, `language`) VALUES ( ?, ?, ?, ?, ?, ?, ?, ?, ?, ? ) ON DUPLICATE KEY UPDATE `channelName`=IF(VALUES(`channelName`) IS NOT NULL, VALUES(`channelName`), `channelName`), `serverID`=IF(VALUES(`serverID`) IS NOT NULL, VALUES(`serverID`), `serverID`), `serverName`=IF(VALUES(`serverName`) IS NOT NULL, VALUES(`serverName`), `serverName`), `region`=IF(VALUES(`region`) IS NOT NULL, VALUES(`region`), `region`), `memberCount`=IF(VALUES(`memberCount`) IS NOT NULL, VALUES(`memberCount`), `memberCount`), `spreadsheet`=IF(VALUES(`spreadsheet`) IS NOT NULL, VALUES(`spreadsheet`), `spreadsheet`), `webhook`=IF(VALUES(`webhook`) IS NOT NULL, VALUES(`webhook`), `webhook`), `modrole`=IF(VALUES(`modrole`) IS NOT NULL, VALUES(`modrole`), IF(`modrole` IS NOT NULL, `modrole`, 'botmods')), `language`=IF(VALUES(`language`) IS NOT NULL, VALUES(`language`), IF(`language` IS NULL, 'ENG_US', `language`));"
-    	
-    }    
-    
+    "translate": {
+      "aliases": [
+        "gt"
+      ],
+      "procedure": "doTranslate",
+      "args": [
+        "id",
+        "lang",
+        "num"
+      ],
+      "help": {
+        "id": "translate",
+        "title": "Google translate",
+        "text": "Quick translate specified number of previous user messages to specified language",
+        "example": "%PREFIX%%COMMAND% <@mention> <lang> <num>"
+      },
+      "subcommands": {}
+    }
+  }
 }
